tasks/pyproject.toml
[tool.poetry]
name = "tasks"
version = "0.1.0"
description = ""
authors = ["Test Name for Update Again <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
black = "^24.3.0"
pre-commit = "^3.6.2"
requests = "^2.31.0"
jupyter = "^1.0.0"
google-cloud = "^0.34.0"
poetry-plugin-export = "^1.7.1"
isort = "^5.13.2"
python-dotenv = "^1.0.1"
functions-framework = "^3.5.0"
google-cloud-storage = "^2.16.0"
pyproj = "^3.6.1"
shapely = "^2.0.3"
google-cloud-bigquery = "^3.20.0"
pyarrow = "^15.0.2"
geojson = "^3.1.0"
flask = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
